refactoring devide file
